+++ ui/src/App.js
@@ -2,7 +2,6 @@ import CssBaseline from '@mui/material/CssBaseline';
 import { DockerMuiThemeProvider } from '@docker/docker-mui-theme';
 import Logo from './Logo';
 import Menu from "./Menu";
-import Data from "./Data";
 import "./App.css";
 
 
@@ -12,7 +11,6 @@ function App() {
       <CssBaseline />
       <Logo />
       <Menu />
-      <Data />
     </DockerMuiThemeProvider>
   );
 }