+++ ui/src/Menu.js
@@ -2,41 +2,40 @@ import { useState } from 'react';
 import { Card, CardHeader, CardContent, IconButton, Tooltip } from "@mui/material";
 import ExpandLessRounded from '@mui/icons-material/ExpandLessRounded';
 import ExpandMoreRounded from '@mui/icons-material/ExpandMoreRounded';
+import MesosTasksIcon from '@mui/icons-material/Apps';
+import MesosFrameworksIcon from '@mui/icons-material/AutoAwesomeMotion';
+import Tasks from "./Tasks";
+import Frameworks from "./Frameworks";
 
 export default function MainMenu() {
-  const [expanded, setExpanded] = useState(false);  
-
-
-  const handleExpand = () => {
-    setExpanded(!expanded);
-  }  
-
+  const [tasks, setTasks] = useState(false);  
+  const [frameworks, setFrameworks] = useState(false);  
+  
   return (
     <>
       <Card>
         <CardHeader
           action={
             <>
-              <Tooltip title={expanded ? "Collapse details" : "Expand details"} placement='bottom-end' >
-                <IconButton
-                  onClick={handleExpand}>
-                  {(expanded) && (
-                    <ExpandLessRounded />
-                  )}
-                  {(!expanded) && (
-                    <ExpandMoreRounded />
-                  )}
+              <Tooltip title="Show Tasks" placement='bottom-end' >
+                <IconButton onClick={() => setTasks(!tasks)}>
+                  <MesosTasksIcon />
+                </IconButton>
+              </Tooltip>
+              <Tooltip title="Show Frameworks" placement='bottom-end' >
+                <IconButton onClick={() => setFrameworks(!frameworks)}>
+                  <MesosFrameworksIcon />
                 </IconButton>
               </Tooltip>
             </>
           }
           title={
-            "Mini Mesos"
+            "Mini Cluster"
           }
         />
-        <CardContent hidden={!expanded} sx={{ paddingTop: "0px" }}>
-        </CardContent>
-      </Card >
+      </Card >    
+     {frameworks && <Frameworks />}
+     {tasks && <Tasks />}
     </>
   );
 }