+++ changelog.md
@@ -1,3 +1,6 @@
 # Master
+- Open port to the mesos agent
+- Add simple webui functionality
 
+# v0.1.0
 - First inn